+++ packages/build-user-config/src/userConfig/swc.js
@@ -32,10 +32,10 @@ module.exports = (config, swcOptions, context, { log, getValue }) => {
           react: reactTransformConfig,
           legacyDecorator: true
         },
-        externalHelpers: true
+        externalHelpers: false
       },
       module: {
-        type: 'commonjs',
+        type: 'es6',
         noInterop: false,
         // webpack will evaluate dynamic import, so there need preserve it
         ignoreDynamic: true