+++ packages/plugin-i18n/src/utils/replaceBasename.ts
@@ -2,14 +2,13 @@ function replaceBasename(pathname: string, basename?: string) {
   if (typeof basename !== 'string') {
     return pathname;
   }
-
   if (basename[0] !== '/') {
     // compatible with no slash. For example: ice -> /ice
     basename = `/${basename}`;
   }
 
   if (pathname.startsWith(basename)) {
-    pathname = pathname.substr(basename.length);
+    pathname = pathname.substring(basename.length);
     if (!pathname.startsWith('/')) {
       pathname = `/${pathname || ''}`;
     }