+++ packages/plugin-react-app/CHANGELOG.md
@@ -5,6 +5,7 @@
 - [feat] optimize runtime when build
 - [chore] bump version of `es-module-lexer`
 - [chore] move runtime's dependencies to `@ice/runtime`
+- [fix] adapt swc rule name
 
 ## 2.1.4